Move lang select to display setttings and cleanup
