+++ jsapp/js/components/processing/sidebar/processingSidebar.tsx
@@ -4,7 +4,6 @@ import singleProcessingStore, {
 } from 'js/components/processing/singleProcessingStore';
 import TransxDisplay from './transxDisplay';
 import SidebarDisplaySettings from 'js/components/processing/sidebar/sidebarDisplaySettings';
-import SidebarLabelsSettings from 'js/components/processing/sidebar/sidebarLabelsSettings.component';
 import type {AssetResponse} from 'jsapp/js/dataInterface';
 import SidebarSubmissionData from 'js/components/processing/sidebar/sidebarSubmissionData';
 import SidebarSubmissionMedia from 'js/components/processing/sidebar/sidebarSubmissionMedia';
