+++ jsapp/js/components/processing/sidebar/sidebarDisplaySettings.tsx
@@ -11,16 +11,21 @@ import ToggleSwitch from 'js/components/common/toggleSwitch';
 import Button from 'js/components/common/button';
 import {AsyncLanguageDisplayLabel} from 'js/components/languages/languagesUtils';
 import type {LanguageCode} from 'js/components/languages/languagesStore';
-import type {AssetContent} from 'js/dataInterface';
 import {getActiveTab} from 'js/components/processing/routes.utils';
 import styles from './sidebarDisplaySettings.module.scss';
 import MultiCheckbox from 'js/components/common/multiCheckbox';
 import type {MultiCheckboxItem} from 'js/components/common/multiCheckbox';
 import cx from 'classnames';
+import KoboSelect from 'js/components/common/koboSelect';
 
 export default function SidebarDisplaySettings() {
   const [store] = useState(() => singleProcessingStore);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
+  const [labelLanguage, setLabelLanguage] = useState<LanguageCode | string>(
+    store.getCurrentlyDisplayedLanguage()
+  );
+
+  const displayedLanguageList = store.getDisplayedLanguagesList();
 
   const activeTab = getActiveTab();
 
@@ -175,6 +180,22 @@ export default function SidebarDisplaySettings() {
           </p>
 
           <ul className={styles.options}>
+            <li className={styles.display}>
+              <KoboSelect
+                label={t('Display labels or XML values?')}
+                name='displayedLanguage'
+                type='outline'
+                size='s'
+                options={displayedLanguageList}
+                selectedOption={labelLanguage}
+                onChange={(languageCode) => {
+                  if (languageCode) {
+                    setLabelLanguage(languageCode);
+                  }
+                }}
+              />
+            </li>
+
             {availableDisplays.map((entry) => {
               const isEnabled = selectedDisplays.includes(entry);
 
@@ -251,6 +272,7 @@ export default function SidebarDisplaySettings() {
               // when the modal is closed.
               resetFieldsSelection();
               setSelectedDisplays(store.getDisplays(activeTab));
+              setLabelLanguage(store.getCurrentlyDisplayedLanguage());
               setIsModalOpen(false);
             }}
           />
@@ -264,6 +286,7 @@ export default function SidebarDisplaySettings() {
             onClick={() => {
               applyFieldsSelection();
               store.setDisplays(activeTab, selectedDisplays);
+              store.setCurrentlyDisplayedLanguage(labelLanguage);
               setIsModalOpen(false);
             }}
           />
