--- jsapp/js/components/processing/sidebar/sidebarLabelsSettings.module.scss
@@ -1,14 +1,0 @@
-@use 'scss/sizes';
-
-$kobo-modal-padding: sizes.$x30;
-
-.root {
-  height: sizes.$x48;
-  display: flex;
-  justify-content: right;
-  align-items: center;
-}
-
-.selectWrapper {
-  padding: $kobo-modal-padding;
-}
